+++ assemblyadapter-list/src/androidTest/java/com/github/panpf/assemblyadapter/list/test/expandable/AssemblyExpandableListAdapterTest.kt
@@ -549,6 +549,66 @@ class AssemblyExpandableListAdapterTest {
         }
     }
 
-    // todo test hasObservers
-    // todo test isChildSelectable
+    @Test
+    fun testMethodHasObservers() {
+        AssemblyExpandableListAdapter<TextGroup, Text>(
+            listOf(TextGroupItemFactory(), TextItemFactory())
+        ).apply {
+            Assert.assertFalse(hasObservers())
+
+            val dataObserver1 = object : DataSetObserver() {}
+            val dataObserver2 = object : DataSetObserver() {}
+
+            registerDataSetObserver(dataObserver1)
+            Assert.assertTrue(hasObservers())
+
+            registerDataSetObserver(dataObserver2)
+            Assert.assertTrue(hasObservers())
+
+            unregisterDataSetObserver(dataObserver1)
+            Assert.assertTrue(hasObservers())
+
+            unregisterDataSetObserver(dataObserver2)
+            Assert.assertFalse(hasObservers())
+        }
+    }
+
+    @Test
+    fun testMethodIsChildSelectable() {
+        AssemblyExpandableListAdapter<TextGroup, Text>(
+            listOf(TextGroupItemFactory(), TextItemFactory()),
+            listOf(
+                TextGroup("hello"),
+                TextGroup("hello", "world", "good"),
+                TextGroup("hello", "bye")
+            )
+        ).apply {
+            Assert.assertNull(childSelectable)
+
+            childSelectable = { adapter, groupPosition, childPosition ->
+                adapter.getChild(groupPosition, childPosition)
+                groupPosition == 1 && childPosition == 2
+            }
+            Assert.assertNotNull(childSelectable)
+
+            assertThrow(IndexOutOfBoundsException::class) {
+                childSelectable(-1, 0)
+            }
+            assertThrow(IndexOutOfBoundsException::class) {
+                childSelectable(0, -1)
+            }
+            Assert.assertFalse(childSelectable(0, 0))
+            assertThrow(IndexOutOfBoundsException::class) {
+                childSelectable(0, 1)
+            }
+            Assert.assertFalse(childSelectable(1, 0))
+            Assert.assertFalse(childSelectable(1, 1))
+            Assert.assertTrue(childSelectable(1, 2))
+            Assert.assertFalse(childSelectable(2, 0))
+            Assert.assertFalse(childSelectable(2, 1))
+            assertThrow(IndexOutOfBoundsException::class) {
+                childSelectable(3, 0)
+            }
+        }
+    }
 }